Fix a build error in 64 bits
In 64bits archs, size_t is 64bits while int is 32 bits.
In fact, the type used for length on the sierialiser is int.

+++ src/server/service/service.cpp
@@ -491,7 +491,7 @@ bool Service::processGetAppGroups(MessageBuffer &buffer, MessageBuffer &send, ui
 
     // success
     Serialization::Serialize(send, SECURITY_MANAGER_API_SUCCESS);
-    Serialization::Serialize(send, gids.size());
+    Serialization::Serialize(send, (int)gids.size());
     for (const auto &gid : gids) {
         Serialization::Serialize(send, gid);
     }
